refactor(sync): move React integration files out of core/ to shared/
`core/` is documented as infrastructure that cannot depend on `features/`
(per CLAUDE.md and check-module-boundaries.sh's allowed-edges table), but
three React integration files in `core/sync/` were importing
`@/features/bin-designer/sync/designAdapter` directly:

  - core/sync/SyncSessionMount.tsx
  - core/sync/useSignOutFlow.tsx
  - core/sync/useDeleteAccountFlow.tsx

The sync engine itself (engine/claim/signOut/deleteAccount/triggers/dialogs)
talks to features only through the `SyncAdapters` interface in
`adapters/types.ts` — exactly the right abstraction. The integration layer
that wires the concrete adapters in was the only thing reaching upward.

Moved the three files to `shared/sync/` where importing from both
`core/sync/*` (engine) and `features/bin-designer/sync/designAdapter`
(implementation) is layering-compliant. Updated the three consumer
imports (App.tsx lazy-load, UserDock, AccountTab) and rewrote the moved
files' relative imports to absolute `@/core/sync/...` paths.

After this, `grep -rln "from '@/features" src/core/` returns zero.

Surfaced by the same architectural integrity audit that produced the
core/units.ts cleanup; one of a series of foundational PRs.
